Use long for hashing the EntryStack

+++ runtime/src/main/java/me/shedaniel/rei/RoughlyEnoughItemsCore.java
@@ -272,7 +272,7 @@ public class RoughlyEnoughItemsCore {
             }
             
             @Override
-            public int hashIgnoreAmount() {
+            public long hashIgnoreAmount() {
                 return getUnwrapped().hashIgnoreAmount();
             }
